Fixes File _crud_handlers.py has 501 lines of code (exceeds 500 allowed).
